Combined all previously created view models into one SettingsViewModel
