+++ src/DAYLY/DAYLY/DAYLY/Views/Settings_DefaultEventDuration.xaml
@@ -6,7 +6,7 @@
              BackgroundColor="{StaticResource AppBackgroundColour}"
              Title="Default event duration">
     <ContentPage.BindingContext>
-        <local:Settings_DefaultEventDurationViewModel/>
+        <local:SettingsViewModel/>
     </ContentPage.BindingContext>
     <StackLayout>
             <Frame Margin="40,50,40,40" CornerRadius="10">
