+++ src/DAYLY/DAYLY/DAYLY/Views/Settings_FirstDayOfWeek.xaml
@@ -6,7 +6,7 @@
              BackgroundColor="{StaticResource AppBackgroundColour}"
              Title="First day of week">
     <ContentPage.BindingContext>
-        <local:Settings_FirstDayOfWeekViewModel/>
+        <local:SettingsViewModel/>
     </ContentPage.BindingContext>
     <StackLayout>
             <Frame Margin="40,50,40,40" CornerRadius="10">
